+++ lib/src/romanizers/japanese.dart
@@ -5,9 +5,7 @@ import 'package:kuromoji/src/tokenizer.dart';
 import 'package:romanize/romanize.dart';
 
 class JapaneseRomanizer extends Romanizer {
-  JapaneseRomanizer() : super(language: 'japanese') {
-    init();
-  }
+  JapaneseRomanizer() : super(language: 'japanese');
 
   static const kanaKit = KanaKit(
     config: KanaKitConfig(