+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [next]
+
+- Japanese Kanji support added using [`kuromoji`](https://pub.dev/packages/kuromoji).
+
 ## 0.0.1
 
 - Initial version.